requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-SocketIO==5.3.6
python-dotenv==1.0.0
Werkzeug==3.0.1
Pillow==10.2.0
requests==2.31.0
gunicorn==21.2.0
gevent==24.2.1
gevent-websocket==0.10.1
simple-websocket==1.0.0
